Blackhole broker implementation
Summary:
A blackhole broker is happy to receive instructions to send message
and do absolutely nothing.

Use cases: prototypes, drop-in replacement in application when there
is any need to disable the broker but it's not convenient to set it
by configuration or a lightweight mock for unit testing.

Test Plan: `phpunit`

Reviewers: dereckson

Differential Revision: http://devcentral.nasqueron.org/D203