Allow to set an exchange as durable
Summary:
The Nasqueron notifications center needs, as explained on T781,
to declare an exchange as durable.

This change allows so in `AMQPBroker::setExchanteTarget`.

This fixes the following `AMQPProtocolChannelException`:
```
PRECONDITION_FAILED - inequivalent arg 'durable' for exchange
'notifications' in vhost 'dev':
received 'false' but current is 'true'
```

Test Plan:
Declare a durable exchange, when an exchange is already declared
in the broker.

Reviewers: dereckson

Maniphest Tasks: T781

Differential Revision: http://devcentral.nasqueron.org/D336